Module mathru::algebra::abstr

Abstract algebra

https://en.wikipedia.org/wiki/Abstract_algebra

Modules

cast

Cast module

Structs

AbsDiff

The requisite parameters for testing for approximate equality using a absolute difference based comparison.

Addition

The addition operator, commonly symbolized by $+$.

Complex

Complex number in cartesian form

Multiplication

The multiplication operator, commonly symbolized by $*$.

Polynomial

Polynomial expression

Relative

The requisite parameters for testing for approximate equality using a relative based comparison.

Traits

AbelianGroup

An Abelian group is a commutative group.

AbelianGroupAdd

AbelianGroupMul

AbsDiffEq

Equality that is defined using the absolute difference of two numbers.

CommutativeRing

Field

Field

Group

A Group is a triple $(\mathbb{M}, \circ, e)$, composed by a set $\mathbb{M}$ and a binary inner operation $\circ$ and the element $e \in \mathbb{M}$

GroupAdd

GroupMul

Identity

A type that is equipped with identity.

Integer

Integer

Loop

A quasigroup with an unique identity element.

Magma

A Magma is a pair $(\mathbb{M}, \circ)$, composed by a set $\mathbb{M}$ and a binary inner operation $\circ$:

MagmaAdd

Syntactic sugar for Magma::::operate(a, b)

MagmaMul

Syntactic sugar for Magma::::operate(a, b)

MatrixMultiply

Monoid

A Monoid is a triple $(\mathbb{M}, \circ, e)$, composed by a set $\mathbb{M}$ and a binary inner operation $\circ$ and the element $e \in \mathbb{M}$

MonoidAdd

MonoidMul

Natural

Natural number

One

Operator

Quasigroup

A quasigroup is a magma which has the divisibility property (or Latin square property). Divisibility is a weak form of right and left invertibility.

Real

Real number

RelativeEq

Equality comparisons between two numbers using both the absolute difference and relative based comparisons.

Ring

Ring

Scalar

comparisons, basic numeric operations, and string conversion.

Semigroup

A Semigroup is a pair $(\mathbb{S}, \circ)$, composed by a set $\mathbb{S}$ and a binary inner operation $\circ$: # Definition

SemigroupAdd

SemigroupMul

Sign

Sign trait

Zero