Struct mathru::algebra::abstr::Polynomial

pub struct Polynomial<T> { /* private fields */ }

Polynomial expression

Implementations

impl<T> Polynomial<T>

pub fn from_coef(coef: Vec<T>) -> Polynomial<T>

Creates a new polynomial with the given coefficients

Arguments

`coef: Coefficients

Panics

If the coef is an empty vector

Example

(3x^2 + 2x + 1)

use mathru::algebra::abstr::Polynomial;

let a: Polynomial<f64> = Polynomial::from_coef(vec![3.0, 2.0, 1.0]);

pub fn from_root(root: Vec<T>) -> Polynomial<T>where
T: Field + Scalar,

Creates a new polynomial from the given roots

https://en.wikipedia.org/wiki/Vieta%27s_formulas

Arguments

root: Roots

Panics

If the root is an empty vector

Example

(x - 1)(x - 2)(x - 3) = -6 + 11x -6x^2 + x^3

use mathru::algebra::abstr::Polynomial;

let a: Polynomial<f64> = Polynomial::from_root(vec![1.0, 2.0, 3.0]);

let b: Polynomial<f64> = Polynomial::from_coef(vec![1.0, -6.0, 11.0, -6.0]);

assert_eq!(b, a);

pub fn from_legendre(n: u32) -> Polynomial<T>where
T: Field + Scalar + AbsDiffEq<Epsilon = T>,

Creates a Legendre polynomial with the given degree

https://en.wikipedia.org/wiki/Legendre_polynomials

Arguments

n: Degree of the Legendre polynomial

Example

use mathru::algebra::abstr::Polynomial;

let p = Polynomial::from_legendre(2);
let p_ref = Polynomial::from_coef(vec![1.5, 0.0, -0.5]);

assert_eq!(p_ref, p)

pub fn from_bernstein(v: u32, n: u32) -> Polynomial<T>where
T: Field + Scalar + Power,

Creates a Bernstein polynomial of degree n

https://en.wikipedia.org/wiki/Bernstein_polynomial

Arguments

n: Degree of the Bernstein polynomial

v:

Panics

if v > n

Example

use mathru::algebra::abstr::Polynomial;

let p = Polynomial::from_bernstein(1, 3);
let p_ref = Polynomial::from_coef(vec![3.0, -6.0, 3.0, 0.0]);

assert_eq!(p_ref, p)

pub fn from_chebyshev_t(n: u32) -> Polynomial<T>where
T: Field + Scalar + AbsDiffEq<Epsilon = T>,

Creates a Chebyshev polynomial of first kind with the given degree

https://en.wikipedia.org/wiki/Chebyshev_polynomials

Arguments

n: Degree of the Chebyshev polynomial

Example

use mathru::algebra::abstr::Polynomial;

let p = Polynomial::from_chebyshev_t(2);
let p_ref = Polynomial::from_coef(vec![2.0, 0.0, -1.0]);

assert_eq!(p_ref, p)

pub fn from_chebyshev_u(n: u32) -> Polynomial<T>where
T: Field + Scalar + AbsDiffEq<Epsilon = T>,

Creates a Chebyshev polynomial of second kind with the given degree

https://en.wikipedia.org/wiki/Chebyshev_polynomials

Arguments

n: Degree of the Chebyshev polynomial

Example

use mathru::algebra::abstr::Polynomial;

let p = Polynomial::from_chebyshev_u(2);
let p_ref = Polynomial::from_coef(vec![4.0, 0.0, -1.0]);

assert_eq!(p_ref, p)

impl<T> Polynomial<T>where
T: Field + Scalar,

pub fn eval(&self, x: T) -> T

Evaluate the polynomial with horner’s rule

Argument

x: The polynomial is evaluated at this value

Example

(1 + 2x + 3x^2) = p(x) \\\\
p(2) = 17

use mathru::algebra::abstr::Polynomial;

let a: Polynomial<f64> = Polynomial::from_coef(vec![3.0, 2.0, 1.0]);

assert_eq!(17.0, a.eval(2.0));

impl<T> Polynomial<T>

pub fn degree(&self) -> usize

Returns the degree of the polynomial

Example

deg(1 + 2x + 3x^2) =2

use mathru::algebra::abstr::Polynomial;

let a: Polynomial<f64> = Polynomial::from_coef(vec![1.0, 2.0, 3.0]);

assert_eq!(2, a.degree())

impl<T> Polynomial<T>where
T: Zero + AbsDiffEq,

pub fn reduce(self) -> Self

impl<T> Polynomial<T>where
T: Field + Scalar + AbsDiffEq<Epsilon = T>,

pub fn differentiate(&self) -> Polynomial<T>

Differentiate polynomial

Example

p(x) = 3x^3 + 5x^2 + 3x + 1

\frac{\partial p(x)}{\partial x} = 9x^2 + 10x + 3

use mathru::algebra::abstr::Polynomial;

let c: Polynomial<f64> = Polynomial::from_coef(vec![3.0, 5.0, 3.0, 1.0]);
let c_s: Polynomial<f64> = Polynomial::from_coef(vec![9.0, 10.0, 3.0]);

assert_eq!(c_s, c.differentiate());

pub fn integrate(&self) -> Polynomial<T>

Integrate polynomial

Example

p(x) = 3x^2 + 2x + 1

\int p(x) dx = c + x + x^2 + x^3

use mathru::algebra::abstr::Polynomial;

let c: Polynomial<f64> = Polynomial::from_coef(vec![3.0, 2.0, 1.0]);
let c_s: Polynomial<f64> = Polynomial::from_coef(vec![1.0, 1.0, 1.0, 0.0]);

assert_eq!(c_s, c.integrate());