Struct mathru::algebra::linear::matrix::matrix_impl::Matrix

pub struct Matrix<T> { /* private fields */ }

Implementations

impl<T> Matrix<T>

pub fn iter(&self) -> MatrixIterator<'_, T>

pub fn iter_mut(&mut self) -> MatrixIteratorMut<'_, T>

pub fn row_into_iter(&self) -> MatrixRowIntoIterator<'_, T>

pub fn column_into_iter(&self) -> MatrixColumnIntoIterator<'_, T>

impl<T> Matrix<T>where
T: Clone,

pub fn apply_mut(self: Matrix<T>, f: &dyn Fn(&T) -> T) -> Matrix<T>

Applies the function f on every element in the matrix

pub fn apply(self: &Matrix<T>, f: &dyn Fn(&T) -> T) -> Matrix<T>

pub fn mut_apply(self: &mut Matrix<T>, f: &dyn Fn(&mut T) -> T)

impl<T> Matrix<T>where
T: Field + Scalar + Power,

pub fn trace(&self) -> Twhere
T: Field + Scalar,

Calculates the trace of a matrix

Arguments

self: square matrix

Panics

if it is not a square matrix

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(2, 2, vec![1.0, -2.0, 3.0, -7.0]);
let tr: f64 = a.trace();

assert_eq!(-6.0, tr);

impl<T> Matrix<T>where
T: Field + Scalar,

pub fn get_column(&self, i: usize) -> Vector<T>

pub fn get_row(&self, i: usize) -> Vector<T>

return row vector

i: row

pub fn set_column(&mut self, column: &Vector<T>, i: usize)

set column

pub fn set_row(&mut self, row: &Vector<T>, i: usize)

set row

Arguments

‘row’

‘i’

Panics

impl<T> Matrix<T>where
T: Field + Scalar + Power,

pub fn givens(m: usize, i: usize, j: usize, c: T, s: T) -> Self

pub fn givens_cosine_sine_pair(a: T, b: T) -> (T, T)

function [c,s ] = Givens(a,b) Givens rotation computation Determines cosine-sine pair (c,s) so that [c s;-s c]’*[a;b] = [r;0] GVL4: Algorithm 5.1.3

impl<T> Matrix<T>where
T: Field + Scalar + Power,

pub fn householder(v: &Vector<T>, k: usize) -> Self

Returns the householder matrix

Arguments

v: Column vector k: index 0 <= k < m

Panics

if index out of bounds

impl<T> Matrix<T>where
T: Field + Scalar,

pub fn get_slice(
&self,
row_s: usize,
row_e: usize,
column_s: usize,
column_e: usize
) -> Matrix<T>

Returns a slice of the matrix

Arugments

0 <= row_s < m
0 <= row_e < m
0 <= column_s < n
0 <= column_e <= n

row_s: start row
row_e: end row
column_s: start column
column_e: end column

Example

use mathru::algebra::linear::{Matrix};

let mut a: Matrix<f64> = matrix![1.0, -2.0; 3.0, -7.0];
a = a.get_slice(0, 0, 1, 1);

let a_ref: Matrix<f64> = Matrix::new(1, 1, vec![-2.0]);

assert_eq!(a_ref, a);

pub fn set_slice(self, slice: &Self, row: usize, column: usize) -> Matrix<T>

Replaces parts of the matrix with the given values

Arugments

0 <= row < m
0 <= column < n

Example

use mathru::algebra::linear::{Matrix};

let mut a: Matrix<f64> = matrix![   1.0, 0.0;
                                    3.0, -7.0];

let b: Matrix<f64> = matrix![2.0, -1.0];
a = a.set_slice(&b, 0, 0);

let a_updated: Matrix<f64> = matrix![   2.0, -1.0;
                                        3.0, -7.0];

assert_eq!(a_updated, a);

impl<T> Matrix<T>where
T: Clone + Copy,

pub fn new(m: usize, n: usize, data: Vec<T>) -> Self

Creates a new Matrix object

Fortran like, column wise

[ 0, 1, 2] 3, 4, 5, 6, 7, 8 ] => vec![ 0, 3, 6, 1, 4, 7, 2, 5, 8]

impl<T> Matrix<T>

pub fn convert_to_vec(self) -> Vec<T>

impl<T> Matrix<T>where
T: Scalar + Clone + Copy,

pub fn new_random(m: usize, n: usize) -> Matrix<T>

impl<T> Matrix<T>where
T: Field + Scalar,

pub fn zero(m: usize, n: usize) -> Self

Returns the zero matrix(additive neutral element)

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(2, 2, vec![1.0, 0.0, 3.0, -7.0]);
let b: Matrix<f64> = &a + &Matrix::zero(2, 2);

impl<T> Matrix<T>where
T: Field + Scalar,

pub fn one(size: usize) -> Self

Returns the eye matrix(multiplicative neutral element)

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(2, 2, vec![1.0, 0.0, 3.0, -7.0]);
let b: Matrix<f64> = &a * &Matrix::one(2);

pub fn ones(m: usize, n: usize) -> Self

impl<T> Matrix<T>where
T: Field + Scalar + Power + AbsDiffEq,

pub fn pinv(&self) -> Result<Matrix<T>, ()>

Calculates the pseudo inverse matrix

A^+ = (A^TA)^-1A^T

impl<T> Matrix<T>

pub fn dim(&self) -> (usize, usize)

Returns the matrix dimension

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(4, 2, vec![1.0, 0.0, 3.0, 0.0, 1.0, -7.0, 0.5, 0.25]);
let (m, n) = a.dim();

assert_eq!(4, m);
assert_eq!(2, n);

pub fn nrows(&self) -> usize

Returns the number of rows

Example

use mathru::algebra::linear::{Matrix};

let a: Matrix<f64> = Matrix::new(4, 2, vec![1.0, 0.0, 3.0, 0.0, 1.0, -7.0, 0.5, 0.25]);
let m: usize = a.nrows();

assert_eq!(4, m);

pub fn ncols(&self) -> usize

Returns the number of columns

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(4, 2, vec![1.0, 0.0, 3.0, 0.0, 1.0, -7.0, 0.5, 0.25]);
let n: usize = a.ncols();

assert_eq!(2, n);

impl<T> Matrix<T>where
T: Field + Scalar + Power + AbsDiffEq<Epsilon = T>,

pub fn dec_eigen(self) -> Result<EigenDec<T>, String>

Computes the eigenvalues of a real matrix

Arguments

Example

use mathru::algebra::linear::{matrix::EigenDec, Matrix, Vector};

let a: Matrix<f64> = Matrix::new(3, 3, vec![1.0, -3.0, 3.0, 3.0, -5.0, 3.0, 6.0, -6.0, 4.0]);
let eigen: EigenDec<f64> = a.dec_eigen().unwrap();

pub fn eigenvalue_r(&self) -> Vector<T>

pub fn eigenvector_r(&self, value: &Vector<T>) -> Matrix<T>

impl<T> Matrix<T>where
T: Field + Scalar + Power,

pub fn dec_hessenberg(&self) -> HessenbergDec<T>

Decomposes self in to the M

q * h * q^T = self

Arguments

Return

(q, h)

Panics

if M is not a square matrix

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(3, 3, vec![1.0, 5.0, 3.0, 1.0, 0.0, -7.0, 3.0, 8.0, 9.0]);
let (q, h): (Matrix<f64>, Matrix<f64>) = a.dec_hessenberg().qh();

impl<T> Matrix<T>where
T: Field + Scalar,

pub fn dec_lu(&self) -> Result<LUDec<T>, ()>

Decomposes the matrix into a upper and a lower matrix

PA = LU

Arguments

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(2, 2, vec![1.0, -2.0, 3.0, -7.0]);

let (l, u, p): (Matrix<f64>, Matrix<f64>, Matrix<f64>) = a.dec_lu().unwrap().lup();

impl<T> Matrix<T>where
T: Field + Scalar + Power + AbsDiffEq,

pub fn dec_qr(&self) -> Result<QRDec<T>, ()>

QR Decomposition with Givens rotations

A = QR
Q is an orthogonal matrix
R is an upper triangular matrix

Panics

if A is not a square matrix

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(2, 2, vec![1.0, -2.0, 3.0, -7.0]);

let (q, r): (Matrix<f64>, Matrix<f64>) = a.dec_qr().unwrap().qr();

impl<T> Matrix<T>where
T: Field + Scalar + AbsDiffEq,

pub fn inv_r(&self) -> Result<Matrix<T>, ()>

impl<T> Matrix<T>where
T: Real,

pub fn dec_cholesky(&self) -> Result<CholeskyDec<T>, String>

Decomposes the symmetric, positive definite quadratic matrix A into a lower triangular matrix L A = L L^T

Arguments

A has to be symmetric and positive definite

Panics

If A is not a quadratic matrix

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = matrix![   2.0, -1.0, 0.0;
                               -1.0, 2.0, -1.0;
                                0.0, -1.0,  2.0];

let l: (Matrix<f64>) = a.dec_cholesky().unwrap().l();

impl<T> Matrix<Complex<T>>where
T: Real,
Complex<T>: Scalar,

pub fn dec_cholesky(&self) -> Result<CholeskyDec<Complex<T>>, ()>

Decomposes the symmetric, positive definite quadratic matrix A into a lower triangular matrix L A = L L^T

Arguments

A has to be symmetric and positive definite

Panics

If A is not a quadratic matrix

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = matrix![   2.0, -1.0, 0.0;
                               -1.0, 2.0, -1.0;
                                0.0, -1.0,  2.0];

let l: (Matrix<f64>) = a.dec_cholesky().unwrap().l();

impl<T> Matrix<T>where
T: Field + Scalar + Power,

pub fn det(&self) -> T

Calculates the determinant

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(2, 2, vec![1.0, -2.0, 3.0, -7.0]);
let det: f64 = a.det();
assert_eq!(det, -1.0)

impl<T> Matrix<T>where
T: Field + Scalar + Power,

pub fn dec_sv(&self) -> (Self, Self, Self)

Computes the singular value decomposition

M = U * S * V*

Return

(u, s, v)

Example

use mathru::algebra::linear::Matrix;

let a: Matrix<f64> = Matrix::new(4,
                                 4,
                                 vec![4.0, 1.0, -2.0, 2.0, 1.0, 2.0, 0.0, -2.0, 0.0, 3.0,
                                      -2.0, 2.0, 2.0, 1.0, -2.0, -1.0]);

let (u, s, v): (Matrix<f64>, Matrix<f64>, Matrix<f64>) = a.dec_sv();

pub fn rot(f: T, g: T) -> (T, T, T)

pub fn householder_bidiag(&self) -> (Self, Self, Self)

self is an m times n matrix with m >= n A = UBV^{T} U \in T^{m \times n} B \in T^{n \times n} V \in T^{n \times n}

impl<T> Matrix<T>

pub fn gcd(m: usize, n: usize) -> usize