Struct mathru::algebra::linear::vector::vector_impl::Vector

pub struct Vector<T> {
    pub data: Matrix<T>,
}

Fields
data: Matrix<T>
Implementations

impl<T> Vector<T>

pub fn iter(&self) -> VectorIterator<'_, T>

pub fn iter_mut(&mut self) -> VectorIteratorMut<'_, T>

impl<T> Vector<T>where
T: Field + Scalar + Power,

pub fn p_norm(&self, p: &T) -> T

Computes the p norm

Arguments

p >= 1.0

Panics

p < 1.0

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_column(vec![1.0, 0.0, 3.0, -2.0]);
let p: f64 = 2.0;
let norm_ref: f64 = a.eucl_norm();
let norm: f64 = a.p_norm(&p);

assert_eq!(norm_ref, norm);

impl<T> Vector<T>

pub fn convert_to_vec(self) -> Vec<T>

impl<T> Vector<T>where
T: Field + Scalar + Power + Exponential,

pub fn eucl_norm(&self) -> T

Computes the euclidean norm

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_column(vec![1.0, 0.0, 3.0, -2.0]);
let norm_ref: f64 = 3.7416573867739413;
let norm: f64 = a.eucl_norm();

assert_eq!(norm_ref, norm);

impl<T> Vector<T>where
T: Clone + Copy,

pub fn new_row(data: Vec<T>) -> Self

Returns a row vector

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_row(vec![1.0, 0.0, 3.0, -2.0]);

pub fn new_column(data: Vec<T>) -> Self

Returns a column vector

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_column(vec![1.0, 0.0, 3.0, -2.0]);

pub fn apply(self: Vector<T>, f: &dyn Fn(&T) -> T) -> Self

impl<T> Vector<T>where
T: Scalar,

pub fn new_row_random(n: usize) -> Self

Returns a row vector initialized with random numbers

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_row_random(4);

pub fn new_column_random(m: usize) -> Self

Returns a column vector initialized with random numbers

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_column_random(4);

impl<T> Vector<T>where
T: Field + Scalar,

pub fn transpose(self) -> Self

Returns the transposed vector

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_column(vec![1.0, 0.0, 3.0, -2.0]);
let b: Vector<f64> = a.transpose();

impl<T> Vector<T>where
T: Field + Scalar,

pub fn dotp(&self, rhs: &Self) -> T

Computes the dot product of two vectors

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_column(vec![1.0, 0.0, 3.0, -2.0]);
let b: Vector<f64> = Vector::new_column(vec![-1.0, 2.0, 3.0, 5.0]);
let dotp_ref: f64 = -2.0;

let dotp: f64 = a.dotp(&b);

assert_eq!(dotp_ref, dotp);

pub fn argmax(&self) -> usize

Find the argmax of the vector.

Returns the index of the largest value in the vector.

Examples

use mathru::algebra::linear::Vector;

let a = Vector::new_column(vec![1.0, 2.0, -3.0, 5.0]);
let idx = a.argmax();
assert_eq!(idx, 3);

pub fn argmin(&self) -> usize

Find the argmin of the vector.

Returns the index of the smallest value in the vector.

Examples

use mathru::algebra::linear::Vector;

let a = Vector::new_column(vec![1.0, -2.0, -6.0, 75.0]);
let b = a.argmin();
assert_eq!(b, 2);

impl<T> Vector<T>where
T: Field + Scalar,

pub fn dyadp(&self, rhs: &Self) -> Matrix<T>

Computes the dyadic product of two vectors

Example

use mathru::algebra::linear::{Matrix, Vector};

let a: Vector<f64> = Vector::new_row(vec![1.0, 0.0, 3.0, -2.0]);
let b: Vector<f64> = Vector::new_column(vec![-1.0, 2.0, 3.0, 5.0]);

let m: Matrix<f64> = a.dyadp(&b);

impl<T> Vector<T>where
T: Field + Scalar + Power,

pub fn reflector(&self) -> Vector<T>

impl<T> Vector<T>where
T: Field + Scalar,

pub fn zero(m: usize) -> Self

Returns the zero vector

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_column(vec![0.0, 0.0, 0.0, 0.0]);
let b: Vector<f64> = Vector::zero(4);

assert_eq!(a, b)

impl<T> Vector<T>where
T: Field + Scalar,

pub fn one(m: usize) -> Self

Returns the one vector

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_column(vec![1.0, 1.0, 1.0, 1.0]);
let b: Vector<f64> = Vector::one(4);

assert_eq!(a, b)

impl<T> Vector<T>

pub fn dim(&self) -> (usize, usize)

Returns the vector dimension

Example

use mathru::algebra::linear::Vector;

let a: Vector<f64> = Vector::new_column(vec![1.0, 2.0, 3.0, 4.0]);
let (m, n): (usize, usize) = a.dim();
assert_eq!(4, m);
assert_eq!(1, n);

impl<T> Vector<T>where
T: Field + Scalar,

pub fn get_slice(&self, s: usize, e: usize) -> Vector<T>

Returns a slice of the vector

Arugments

0 <= s < m
0 <= e < m

s: start
e: end

Panics

iff s and e are out of bounds

Example

use mathru::algebra::linear::Vector;

let mut a: Vector<f64> = Vector::new_column(vec![1.0, -2.0, 3.0, -7.0]);
a = a.get_slice(1, 2);

let a_ref: Vector<f64> = Vector::new_column(vec![-2.0, 3.0]);

assert_eq!(a_ref, a);

pub fn set_slice(&mut self, rhs: &Self, s: usize)

Overwrite a slice of the vector with the given values

Arugments

0 <= s < m

s: start

Example

use mathru::algebra::linear::Vector;

let mut a: Vector<f64> = Vector::new_column(vec![1.0, -2.0, 3.0, -7.0]);
let b: Vector<f64> = Vector::new_column(vec![-5.0, 4.0]);
a.set_slice(&b, 1);

let a_ref: Vector<f64> = Vector::new_column(vec![1.0, -5.0, 4.0, -7.0]);

assert_eq!(a_ref, a);