Trait mathru::statistics::distrib::Continuous

pub trait Continuous<T>where
    T: Real,{
    fn pdf(&self, x: T) -> T;
    fn cdf(&self, x: T) -> T;
    fn quantile(&self, p: T) -> T;
    fn mean(&self) -> T;
    fn variance(&self) -> T;
    fn skewness(&self) -> T;
    fn median(&self) -> T;
    fn entropy(&self) -> T;
}

Continuous distribution

Required Methods

fn pdf(&self, x: T) -> T

Probability density function

Arguments

*x:

fn cdf(&self, x: T) -> T

Cumulative distribution function

Arguments

*x:

fn quantile(&self, p: T) -> T

Quantile function, inverse cdf

fn mean(&self) -> T

Mean

fn variance(&self) -> T

Variance

fn skewness(&self) -> T

Skewness is a measure of the asymmetry of the probability distribution of a real-valued random variable about its mean

fn median(&self) -> T

Median is the value separating the higher half from the lower half of a probability distribution.

fn entropy(&self) -> T